pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usgs-quake-stream"
version = "0.1.0"
description = "Real-time USGS earthquake data pipeline using Kafka"
requires-python = ">=3.9"
dependencies = [
    "confluent-kafka>=2.3.0",
    "click>=8.1.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "psycopg2-binary>=2.9.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pydeck>=0.8.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-httpx>=0.28.0",
    "ruff>=0.1.0",
]

[project.scripts]
quake = "quake_stream.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
